Revert "ci: Built glycin-free gdk-pixbuf"
GdkPixbuf now defaults to glycin-based loaders that use bwrap to
sandbox the actual image loading. Unfortunately this broke when
running under gnome-desktop-testing-runner in Docker.

So far we were working around this by disabling the glycin loader,
but at the cost of support for SVG images.

But now glycin merged a fix to skip sandboxing where bwrap fails,
so pull that in to have both working tests and SVG support.

This reverts commit f72fcc499dfd8390b57d6232fe6d72ed5c3b2ddd.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4666>