window: Return result rect in move_resize_internal()
The result rect is the resulting rectangle after applying resize
direction logic, constraints, and other window management logic. It's
the rectangle that Wayland clients are configured to look like, and what
X11 windows are moved and resized to be.

It can be useful to know the expected end result in certain situations,
for example where to warp the mouse cursor to during keyboard driven
resize drags.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4673>