window-drag: Maintain a 'drag' rectangle
This is meant to be the rect that the window is configured to be. Having
such a rect is useful to calculate an up to date window size without
relying on pointer coordinates, for example when performing a keyboard
based window drag.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4673>