backend: Set keymap using new MetaKeymapDescription type
This allows moving some logic to a type that centralizes handling of
describing a keyboard map. It also allows attaching more metadata to the
keymap which will be useful in the future.

The MetaKeymapDescription type is a immutable description of a XKB keyboard
map. Switching related state, i.e. layout index, is done externally to
it.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4699>