window: Add a map inhibit API
A compositor (such as GNOME Kiosk) may need to prevent windows from
showing on screen for various reasons, yet there is no appropriate API
to tell mutter that a window that could be mapped on screen should not
actually show up.

To solve that problem, add a new API that allows to caller to inhibit
mapping windows.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4692>