window: Do not unqueue wayland window for move-resize when acking
If we had queued a window for move-resize updates and some event causes us
to process a previous configuration ack, we don't want to remove the window
from the queue and make it lose the next configuration update.

This fixes the bug where maximized windows may extend under a panel when
locking/unlocking or changing the display scale, where we get multiple struts
updates in quick succession.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/1627
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4695>