keymap-description: Add way to 'lock' a keymap description
This should be interpreted by other entities that tries to set the
keymap to avoid doing so. It's meant to allow a remote desktop service
to be in control of the current keymap and keyboard layout.

This introduces a concept of keymap description owner. Only the same
owner can set a new keymap if a current keymap description is locked.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4699>