remote-desktop: Allow remote desktop servers to set the keyboard layout
This allows remote desktop servers to take control of the keyboard
layout by setting their own keymap. This is meant to for example allow
using the same keymap in the remote system as in the system where the
remote desktop client is running.

For the time being, if the remote desktop server sets the keymap with a
'locked' property, it currently breaks the usage of input methods, as
they currently set what keyboard layout they work with, meaning that if
they can't override the keyboard layout, they wont work. This is a known
issue and deliberately not solved right now.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4699>