mtk: Ensure to get cancellable before stealing task
The execution order of arguments in a function call are pretty much
up to the compiler and/or architecture. This means the order that we
get the task cancellable, and steal it to give ownership to the async
func is pretty much not guaranteed. Get the cancellable first, so that
the order is well defined.

CID: #1657846
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4697>