wayland: Do not handle WINDOW_STATE_NONE in session restore
This matches the paths above, where we leave a NULL rect, this might
trigger crashes down the line if used as-is. Since this state should
never be saved, double check we do nothing with it by bailing out
early, too.

CID: #1657444
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4697>