wayland: Do not force pointer focus on popups
We currently force switch the pointer focus to be the popup
surface whenever a xdg_popup is created, drop this for three
reasons:

- The protocol says "During a popup grab, the client owning the
  grab will receive pointer and touch events for all their surfaces
  as normal". That also means receiving a wl_pointer.enter event iff
  the surface is under the pointer.
- The popup surface actor will be on its way to having its size
  allocated in order to being shown. When we switch focus here, the
  coordinates we can obtain for the actor in this state will be
  nonsensical. We do indeed emit wl_pointer.enter with out of whack
  coordinates when a xdg_popup is shown.
- The reason to add this was
  https://gitlab.gnome.org/GNOME/gtk/-/issues/7414, but this issue
  does not reproduce anymore.

Dropping this, we let the MetaWaylandPointer update itself if
the pointer ends up over any other another client surface, be it
through motion or actor relayouts. This is truer to the xdg-shell
spec, and results in .enter events having correct coordinates.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4703>