wayland: Change implicit pointer grab behavior
In order to honor a bit better the "owner_events lookalike" spirit of
the protocol in describing the interaction of popup grabs with client
events, make the implicit grab surface only effective outside of the
client. Inside the client, enter and leave events will be emitted as
normal between child surfaces.

This is a behavioral change that requires changes to the
popup-focus.metatest stacking test. While at it, test a bit further
the behavior inside popups (checking that focus changes to these), and
outside the implicit grab client (checking that focus returns to the
implicit grab surface).

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4703>