core/launch-context: Implement get_startup_notify_id with nullable AppInfo
As per GNOME/glib@6210708f the get_startup_notify_id vfunc has a
nullable appId parameter, this is because the value is irrelevant in a
wayland-only and XDG activation token world.

However, mutter still tries to use StartupNotification protocol if
possible and this is causing now a crash when activating a notification
since GNOME/gnome-shell@f85ee952.

So, if no appId is provided, let's just provide the minimum to the
StartupNotification protocol so that X11 applications can still "remove"
the launch request, and mutter still uniquely identify them.

This is possible because libSN uses sequential IDs for sequence numbers,
so the ID will be always different.

At the same time we need to pass an empty (not NULL) value as launchee
to sn_launcher_context_initiate(), since that strlen's it and it's not
nullable, even though the value can be unset, being it only for
debugging purposes

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4705>