backends/x11: Create distinct sprites for touch events
The X11 touch support was left in a subpar state after the switch
to ClutterSprite, as the touchpoints are not tracked at all. This
likely broke all multi-touch gestures, beyond single-touch interaction.

To fix this, actually track all additional touchpoints with their
own ClutterSprite. We still reuse the pointer sprite for the
touchpoint that is in charge of emulating pointer events.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4706>