window/wayland: Don't unset last acked config
When a client sporadically resizes, e.g. due to content changing size,
there is no "acked configuration" when applying the window state, but we
still attempted to update the "last acked configuration" to what was
currently acked, which was none. This effectively cleared the last acked
configuration, which wasn't intended.

This caused state changes to revert to the previous window size, since
it compared the last acked configuration to the last sent configuration
to see if what state was up to date.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4391
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4712>