core: Avoid processing twice a same key event during keybindings
The changes in commit c0fd2d9556 to let wayland clients follow more
closely the pressed modifiers resulted in combinations of these special
modifiers with other keys maybe going through process_event() twice
in MetaKeybindings, once when attempted as a fallback when detecting
this situation, and then again because that commit now lets these
events through. This seems to result in inconsistent state and
seemingly repeated actions.

Since we do have good reasons to let these events maybe go past
keybindings (e.g. the keycombo not being managed by the compositor
after all, but the app), it sounds better to let the single
process_event() call happen through the fallbacks. This again results
in the action being handled once.

In X11, we used to replay events in those paths, as we freeze event
processing first when receiving the special modifier key. This has
been shuffled, so that the intent is preserved, and events get
replayed if they were not meant for keybindings.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4368
Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4404
Fixes: c0fd2d9556 ("core: Let key presses of special modifiers through")
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4724>