mdk: Allow configuring size via gdctl by default
Pass a set of possible modes when creating the virtual monitors, if the
monitor is non-resizable. This allows the user to configure the modes of
the devkit monitor using e.g. gdctl. Flipping the resizable switch
recreates the monitor with a dynamic size determined during PipeWire
stream negotation.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4727>