mdk/stream: Split up constructor
One for resizable, and one for non-resizable. The default width for the
resizable is folded into the constructor, and for non-resizable, the
default size was not used anyway, so it could be dropped.

This also fixes an issue where the default passed size used for the
place holder paintable would not have the same size as the eventually
configured virtual CRTC / stream.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4727>