wayland: Confine implicit grab inter-surface crossing within toplevel
Commit d424771439 attempted to make the implicit grab behavior able to cater
for click/drag/release patterns to select items in menus. For this we need
to keep the pointer able to cross between surfaces despite the implicit grab
on the parent surface of the popup.

However this check breaks Xwayland expectations that the implicit grab must
stay in the same window, breaking some clients like xterm due to the unexpected
crossing events.

In order to keep our cake and eat it, make the implicit grab ability to cross
between surfaces apply within the surface tree of a toplevel, instead of
across any surface of the same client.

Notably, menus in Xwayland are not driven by xdg_popup, and are seen as
toplevels themselves. This allows us to keep Xwayland happy, and to keep a
sensible implicit grab behavior without having to think yet again from scratch
how the click/drag/release pattern could ever work without pointer crossing.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4409
Fixes: d424771439 ("wayland: Change implicit pointer grab behavior")
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4729>