clutter: Make internals aware of ClutterCursor
There's a few closely related changes here. The ClutterActor API gained
a vmethod to query a ClutterCursor for a given sprite, with its default
implementation looking up the appropriate ClutterCursor for the given actor
ClutterCursorType. There is also a new method to trigger an invalidation
of this cursor, e.g. in case it changed while a pointer is over it.

This relies on new ClutterBackend API (+vmethod) to resolve a
ClutterCursorType to a ClutterCursor, which will be preserved in the
ClutterSprite. The cursor will be invalidated when transitioning across
actors, or if triggered through the ClutterActor API.

This provides us of escape hatches for ClutterActor implementations to
lean on special implementations of ClutterCursor (useful for Wayland),
while keeping the user-level Clutter API oriented to the ClutterCursorType
enum.

At the moment, this is hooked nowhere, and the ClutterSprite/ClutterBackend
vfuncs are left unimplemented. Also, at the moment the cursor lookup will
end up on a NULL cursor if no actor up the stage provided one, later commits
will flip the switch in making Clutter in charge of the default cursor.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4745>