compositor: Hook to global cursor override in backend
Currently, the only legit use of the root cursor in MetaCursorTracker
is startup notification cursor feedback, as something that happens
anywhere on the stage as opposed to a certain actor.

Use the MetaBackend::override-cursor signal for this, and have a
single point where we can define a global policy. This root cursor
never applied previously to tablet cursors, since their renderers
are not governed by the cursor tracker. This cursor is truly global
so the issue is now moot.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4745>