backends/native: Make ClutterSprites swap cursor renderers
Implement the mechanism that allows cursor renderers to change to
the currently used sprite. This allows input devices other than the
mouse (e.g. tablets) to get the native cursor renderer using a
HW plane.

This preference sticks as long as cursor should be shown on-screen,
either a pointer with cursor inhibition deactivated or a tablet in
proximity. As soon as any of those conditions are not met, the
HW plane is left vacant for any other sprite to make use of it next.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4745>