screen-cast: Add support for device ID negotiation
Device ID negotation allows for the source and sink to negotiate what
DRM device should be used for DMA buffers. For mutter in particular this
currently still means only a single DRM device is advertised, but it
also allows for the consumer of the stream to find out the DRM device
associated with the DMA buffers without guessing or going via some side
channel.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4767>