cogl: Create a RendererEgl
Currently it uses RendererEgl instead of RendererEGL as there is a
struct with that name and we want to move those fields from there to
RendererEGL eventually.

The type is also derivable, so we can eventually derive it in
MetaRendererNative impl and move various Winsys vfuncs to Renderer as
that is where they belong.