window: Add external constraints support
This allows to keep a list of external constraints with the window.

The caller can add/remove constraint objects to each window.

This is preparation work for the following commit that will call into
this to apply external window constraints along the rest of the
constraints.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4692>