backends/native: Fix mapping of mouse/lens tool buttons to actions
These actions are typically mapped in a way that makes sense for
styli, where the tip is the button1/primary/default action, and the
extra buttons are mapped to right/middle actions (and so forth).

This mapping however is non-sensical for mouse and lens tool types,
since those are mouse-alike, and the actions for middle and right
buttons become swapped. These tools do not have further extra buttons,
so it does not make sense to map either the tertiary action, the
first button is always left at its default action.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4414
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4781>