backends/native: Fallback to SW cursor through return value (again)
There may be several other cursor renderers that should not be affected
by the native renderer being able to use the HW cursor or not. We need this
to apply only to the MetaOverlay relative to the cursor, so go back to using
the boolean value really meaning "needs a stage overlay".

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4745>