clutter/color-state-params: Use luminance ratios in needs_tone_mapping
What's relevant for tone mapping isn't the absolute maximum mastering
luminance but the ratio between that and the reference luminance.

This avoids spurious tone mapping in cases where the two luminances are
equal, which can't work correctly (there's a divisor which becomes 0 in
this case).

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4506

v2:
* Adjust maximum mastering luminance of HDR blending color state to
  avoid spurious tone mapping when transforming from blending to output
  color state.
v3:
* Tweak trivial comparisons to cover HDR (and hypothetical source color
  states with reference luminance > maximum mastering luminance).

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4804>