cogl/framebuffer: Add helper to check blit format preconditions
We can only blit if the alpha channel will be something predictable, so
allow to check whether blitting will fail due to this condition or not.
This allows avoiding handling an error which isn't really one, and
directly fall back to drawing if needed.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4805>