output-kms: Force sysfs backlights to 100% when a ref-white is used
We still don't know if a sysfs backlight functions when we switch into
HDR mode where we use the ref-white based backlight. Unfortunately that
also means that it *can* function which limits the ref-white backlight
to a maximum of the sysfs backlight. So let's just max out the sysfs
backlight in that case for now.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4814>