backends/input-settings: Hook up stylus eraser button configuration
libinput 1.29 and later allow configuring an eraser button as a normal
button (instead of the firmware-hardcoded pretend "flip pen around and
use as eraser" whenever the button is pressed). The libinput
configuration takes a mode and a button to send but currently limits
the configurable buttons to BTN_STYLUS through to BTN_STYLUS3. This is
fixed upstream in the (as yet unreleased) libinput 1.32.

We need to implement this a bit differently to be able to distinguish
between the configured eraser button and a normal stylus button - the
eraser button is always configured to send BTN_STYLUS3 and we
intercept that and map it just like the other stylus settings.
No current pen has three buttons and an eraser button anyway (though
a pen with three buttons and an eraser *tip* exists with the Huion
PW600).

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4819>