clutter: Add private API to set "chrome" for grabs
This call will allow for parts of the actor hierarchy to work
in the presence of grabs, despite their position in the actor
tree.

One important usecase is the on-screen keyboard, which is meant
to work on a variety of situations, including grabbing ones like
modal dialogs and the looking glass shell. Currently, these places
manually funnel the events through clutter_actor_event() which
kinda sorta works but requires JS code to manipulate the individual
ClutterEvents. This is a replacement to set this up once, and have
it work without further intervention.

Since the amount of usecases currently sum up to one, it does not
seem to make sense to overcomplicate the API, thus a single actor
is allowed to be set. This will be used in later commits for the
on-screen keyboard solely.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/5171>