backends: Drop API to finish a touch sequence
This was only necessary in the X11 backend, in order to finish
the passive touch grab(s) set up by the compositor, and either
claim full control or give up handling in favor of the app
handling the events.

There's no passive grab mechanisms to cater for in the native
backend, so this API is moot.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4829>