core: Drop meta_display_is_pointer_emulating_sequence()
This was a stopgap solution to retain some sensible behavior for
single-touch so it was consistent on both X11 and Wayland. The
consistency point is moot, and the current infrastructure (passive
and active grab handling in clutter, the gesture framework) make it
less necessary to the point it's already unused in GNOME Shell.

Use the opportunity to drop this API.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4829>